# reimplement string with shared_ptr instead of concrete macro