@jingkaimori 2024-05-21
Because of unnecessary spin lock, std::shared_ptr will not be used in lolly, see #326